backend/requirements.txt
google-genai>=1.0.0
firebase-admin>=6.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
